refactor: extract renderer main boundaries
