+++ packages/insomnia/src/common/path-with-params.ts
@@ -1,0 +1,3 @@
+const VARIABLE_SEARCH_VALUE = /{([^}]+)}/g;
+
+export const pathWithParamsAsPathParameters = (path?: string) => path?.replace(VARIABLE_SEARCH_VALUE, ':$1') ?? '';
